backend/pyproject.toml
[tool.poetry]
name = "ymm-backend"
version = "0.1.0"
description = "YMM Automation Backend"
authors = ["YMM"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.3"
psycopg = "^3.1"
dj-database-url = "^2.1"
celery = "^5.3"
redis = "^5.0"
boto3 = "^1.34"
django-storages = "^1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
